Adds credibility indicator to attendance overview
Adds a visual cue to the attendance overview to indicate the credibility of a user's attendance commitment.
This is achieved by introducing a `credible` prop to the Zusage component, which is then passed down to the Check component.
The Check component now uses a different fill color for the circle icon based on the `credible` prop value.